feat: Add ability to play first song while playlist is processing
This commit adds the ability to play the first song of a playlist while the rest of the songs are still being processed. The logic of adding a song to the playlist has been moved to a separate class, and the relevant methods in the downloader have been changed to static methods.